requirements.txt
pillow>=10.2.0
pyqt5 >= 5.15.10
